BSB: avoid integer overflow. Fixes https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=1963. Credit to OSS Fuzz
git-svn-id: https://svn.osgeo.org/gdal/trunk@38712 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/bsb/bsb_read.c
@@ -788,7 +788,9 @@ int BSBReadScanline( BSBInfo *psInfo, int nScanline,
             while( (byNext & 0x80) != 0 && !bErrorFlag)
             {
                 byNext = BSBGetc( psInfo, psInfo->bNO1, &bErrorFlag );
-                nRunCount = nRunCount * 128 + (byNext & 0x7f);
+                /* Cast to unsigned to avoid int overflow. Even if the */
+                /* value is crazy, we validate it afterwards */
+                nRunCount = (int)((unsigned)nRunCount * 128 + (byNext & 0x7f));
             }
 
             /* Prevent over-run of line data */
